requirements.txt
google.generativeai
chromadb
soundfile
gradio
